Several changes
1. Added dev-dependencies to package.json and updated dependencies to
use racer instead of derby. As a consequence made required changes in
index.js.
2. Moved helper functions to a separate file (helper_functions.js) and
included it in test/view.js and made necessary changes in view.js.
3. Updated addData() in helper_functions.js to not return the model.
