+++ package.json
@@ -7,7 +7,11 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
-    "derby": "^0.6.0-alpha31"
+    "racer": "0.6.0-alpha45"
+  },
+   "devDependencies": {
+    "expect.js": "~0.3.1",
+    "mocha": "~1.17.1"
   },
   "author": "cjblomqvist <carl@cjblomqvist.com>",
   "license": "MIT",
