+++ test/view.js
@@ -1,39 +1,9 @@
-var expect    = require('../node_modules/expect.js');
-var Model     = require('../node_modules/racer/lib/Model');
+var expect    = require('expect.js');
+var Model     = require('racer/lib/Model');
+var helper    = require('./helper_functions.js');
 require('../index.js')({Model: Model});
 
-/* ----------- HELPER FUNCTIONS ----------- */
-// Adds data to the model
-function addData(model) {
-  var fruits = [  {name: 'apple',  color: 'red', amount: 5, id: 'appleId'},
-                  {name: 'orange', color: 'orange', amount: 10, id: 'orangeId'},
-                  {name: 'banana', color: 'yellow', amount: 15, id: 'bananaId'},
-                  {name: 'lemon',  color: 'yellow', amount: 20, id: 'lemonId'} ];
 
-  for (var i = 0, len = fruits.length; i < len; i++) {
-    model.add('fruits', fruits[i]);
-  }
-  return model;
-}
-
-
-// Returns an object containing key-object pairs for the specified ids
-function createExpectedResult(model, idArray) {
-  var result = {};
-  var fruitObject;
-  var key;
-
-  for(var i = 0, len = idArray.length; i < len; i++) {
-    fruitObject = model.get('fruits.'+ idArray[i]); 
-    key = fruitObject.name + '*' + fruitObject.color + '*' + fruitObject.amount*2;
-    result[key] = fruitObject;
-  }
-  return result;
-}
-/* ----------- END HELPER FUNCTIONS ----------- */
-
-
-/* ----------- TESTS ----------- */
 describe('Derby-View', function() {  
   describe('Setting up view', function() {
     it('empty collection: returns name of created view', function() {
@@ -51,7 +21,7 @@ describe('Derby-View', function() {
     it('non-empty collection: returns name of created view', function() {
       var model    = (new Model).at('_page');
       
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color, '_page.fruits.' + fruit.id);
@@ -79,7 +49,7 @@ describe('Derby-View', function() {
     it('non-empty collection: adds view containing results to the model', function() {
       var model    = (new Model).at('_page');
 
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -88,7 +58,7 @@ describe('Derby-View', function() {
 
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits'); // Reference view with non-empty collection
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
   });
@@ -105,7 +75,6 @@ describe('Derby-View', function() {
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
 
-      // Add new item to empty collection
       model.add('fruits', {name: 'grapefruit', color: 'orange', amount: 10, id: 'grapefruitId'});   
       expect(model.get('filteredFruits')).to.be.empty();
     });
@@ -123,14 +92,14 @@ describe('Derby-View', function() {
 
       // Add new item to empty collection
       model.add('fruits', {name: 'banana', color: 'yellow', amount: 15, id: 'bananaId'});
-      var expectedFruits = createExpectedResult(model, ['bananaId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
 
     it('non-empty colection: does NOT add to view as conditions are NOT satisfied', function() {
       var model    = (new Model).at('_page'); 
 
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -141,14 +110,14 @@ describe('Derby-View', function() {
 
       // Add new item to non-empty collection
       model.add('fruits', {name: 'grapefruit', color: 'orange', amount: 10, id: 'grapefruitId'});
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });  
 
     it('non-empty colection: adds to view as conditions ARE satisfied', function() {
       var model    = (new Model).at('_page');
 
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -159,7 +128,7 @@ describe('Derby-View', function() {
 
       // Add new item to non-empty collection
       model.add('fruits', {name: 'mango', color: 'yellow', amount: 15, id: 'mangoId'});      
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId', 'mangoId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId', 'mangoId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
   });
@@ -168,7 +137,7 @@ describe('Derby-View', function() {
     it('removes item from view as it was included', function() {
       var model    = (new Model).at('_page');
   
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -178,20 +147,20 @@ describe('Derby-View', function() {
       // Before removing included item
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
       
       model.del('fruits.bananaId'); // Remove included item
             
       // After removing included item
-      expectedFruits = createExpectedResult(model, ['lemonId']);
+      expectedFruits = helper.createExpectedResult(model, ['lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
 
     it('does NOT make any changes to view as item was NOT included', function() {
       var model    = (new Model).at('_page');
 
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -201,13 +170,13 @@ describe('Derby-View', function() {
       // Before removing item NOT included in view
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
 
       model.del('fruits.appleId'); // Remove item
 
       // After removing item NOT included in view
-      expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });    
   }); 
@@ -216,7 +185,7 @@ describe('Derby-View', function() {
      it('adds updated item to view as conditions ARE now satisfied', function() {
       var model    = (new Model).at('_page');
       
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -226,7 +195,7 @@ describe('Derby-View', function() {
       // Before updating item
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
 
       model.set('fruits.appleId.color', 'yellow');  // Update item
@@ -234,14 +203,14 @@ describe('Derby-View', function() {
       // After updating, item is now included in the view
       view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      expectedFruits = createExpectedResult(model, ['appleId', 'bananaId', 'lemonId']);
+      expectedFruits = helper.createExpectedResult(model, ['appleId', 'bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
 
     it('removes item from view as conditions are NO longer satisfied', function() {
       var model    = (new Model).at('_page');
       
-      model = addData(model); // Add items to collection
+      helper.addData(model); // Add items to collection
       model.fn('yellowFruits', function(emit, fruit) {
         if (fruit.color === 'yellow') {
           emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -251,7 +220,7 @@ describe('Derby-View', function() {
       // Before updating item included in view
       var view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      var expectedFruits = createExpectedResult(model, ['bananaId', 'lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['bananaId', 'lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
      
       model.set('fruits.bananaId.color', 'green'); // Update item
@@ -259,7 +228,7 @@ describe('Derby-View', function() {
       // After updating the item, it is now excluded from the view
       view = model.at('fruits').view('yellowFruits');
       view.ref('_page.filteredFruits');
-      var expectedFruits = createExpectedResult(model, ['lemonId']);
+      var expectedFruits = helper.createExpectedResult(model, ['lemonId']);
       expect(model.get('filteredFruits')).to.eql(expectedFruits);
     });
   });
@@ -316,7 +285,7 @@ describe('Derby-View: Events', function() {
       it('Item included in view: triggers "change" event', function() {
         var model    = (new Model).at('_page');
 
-        model = addData(model); // Add items to collection
+        helper.addData(model); // Add items to collection
         model.fn('yellowFruits', function(emit, fruit) {
           if (fruit.color === 'yellow') {
             emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -335,7 +304,7 @@ describe('Derby-View: Events', function() {
       it('Item NOT included in view: triggers "change" event', function() {
         var model    = (new Model).at('_page');
 
-        model = addData(model); // Add items to collection
+        helper. addData(model); // Add items to collection
         model.fn('yellowFruits', function(emit, fruit) {
           if (fruit.color === 'yellow') {
             emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -356,7 +325,7 @@ describe('Derby-View: Events', function() {
       it('Item included in view: triggers "change" event', function() {
         var model    = (new Model).at('_page');
 
-        model = addData(model); // Add items to collection
+        helper.addData(model); // Add items to collection
         model.fn('yellowFruits', function(emit, fruit) {
           if (fruit.color === 'yellow') {
             emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -374,9 +343,9 @@ describe('Derby-View: Events', function() {
       });
 
       it('Item NOT included in view: triggers "change" event', function() {
-         var model    = (new Model).at('_page');
+        var model    = (new Model).at('_page');
 
-        model = addData(model); // Add items to collection
+        helper.addData(model); // Add items to collection
         model.fn('yellowFruits', function(emit, fruit) {
           if (fruit.color === 'yellow') {
             emit(fruit.name + '*' + fruit.color + '*' + fruit.amount*2, '_page.fruits.' + fruit.id);
@@ -394,4 +363,3 @@ describe('Derby-View: Events', function() {
       });
     });
   });
-/* ----------- END TESTS ----------- */
